+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v1.0.0
+
+18.12.2024.
+
+### Changed
+
+- **Breaking Change**: Removed `thumbWidth` property; solved it more elegantly using CSS calc()
+- Added React component
+
 ## v0.9.10
 
 08.12.2024.